--- src/main/java/sws/songpin/entity/member/dto/MemberRequestDto.java
@@ -1,4 +1,0 @@
-package sws.songpin.entity.member.dto;
-
-public record MemberRequestDto() {
-}
