+++ src/main/java/sws/songpin/entity/member/dto/request/MemberRequestDto.java
@@ -1,0 +1,4 @@
+package sws.songpin.entity.member.dto.request;
+
+public record MemberRequestDto() {
+}
